refactor: split console-commands.js into subdirectory modules
Breaks the 1039-line flat file into six focused modules:

  registry.js    — CommandDef typedef, Map, registerCommand/getCommand
  completions.js — tab-completion providers, getRevealedAliases, tabComplete
  resolvers.js   — resolveNode/Card/ImplicitNode, dispatch, resolveWanAccess
  cmd-status.js  — all cmdStatus* display functions
  commands.js    — all CommandDef objects (COMMANDS array)
  index.js       — entry point, registers commands, re-exports public API

Tests for each surface moved into the same directory and renamed to match
their corresponding module (tab-complete → completions, console-commands → commands).
